+++ frontend/app/src/entities/events/ui/branch-events/branch-event.tsx
