+++ frontend/app/src/entities/events/ui/branch-events/global-branch-event.tsx
