+++ frontend/app/src/entities/events/ui/global-event.tsx
@@ -5,10 +5,10 @@ import { classNames } from "@/shared/utils/common";
 import { Icon } from "@iconify-icon/react";
 import { format } from "date-fns";
 import { BRANCH_EVENTS, GROUP_EVENTS, STANDARD_EVENTS } from "../constants";
-import { BranchEvent } from "./global-branch-event";
-import { GroupEvent } from "./global-group-event";
-import { NodeEvent } from "./global-node-event";
+import { BranchEvent } from "./branch-events/global-branch-event";
 import { StandardEvent } from "./global-standard-event";
+import { GroupEvent } from "./group-events/global-group-event";
+import { NodeEvent } from "./node-events/global-node-event";
 
 const GlobalEventDisplay = ({ __typename, ...props }: EventType) => {
   if ("attributes" in props) {
