+++ frontend/app/src/entities/events/ui/group-events/global-group-event.tsx
