+++ frontend/app/src/entities/events/ui/group-events/group-event.tsx
