+++ frontend/app/src/entities/events/ui/node-events/constants.ts
@@ -1,0 +1,5 @@
+export const NODE_EVENTS_MAPPING: Record<string, string> = {
+  "infrahub.node.created": "created",
+  "infrahub.node.updated": "updated",
+  "infrahub.node.deleted": "deleted",
+};
